refactor: factor out find_all_related_roles
